feat-#16: move events to domain
