+++ src/item/domain/events/handlers/item-deleted.handler.ts
