+++ src/item/domain/events/handlers/item-owner-notified.handler.ts
