+++ src/item/domain/events/impl/item-created.event.ts
@@ -1,6 +1,6 @@
 import { IEvent } from '@nestjs/cqrs';
 
-import { CreateItemDto } from '../../dtos';
+import { CreateItemDto } from '../../../dtos';
 
 export class ItemCreatedEvent implements IEvent {
   constructor(public readonly createItemDto: CreateItemDto) {}
