+++ src/item/domain/events/impl/item-deleted.event.ts
